Split out cli
